backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-generativeai 
docxtpl
docx2pdf
python-multipart
pypdf
docx
python-docx
googlesearch-python
google-api-python-client
jsonschema
beautifulsoup4
mysql-connector-python
bcrypt
pydantic[email]
fastapi-mail